% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footprintsScanner.R
\name{DB}
\alias{DB}
\title{helper function for differential binding}
\usage{
DB(counts, libSize, group, default.bcv = 0.3)
}
\arguments{
\item{counts}{count table}

\item{libSize}{library size}

\item{group}{group design}

\item{default.bcv}{a reasonable dispersion value}
}
\value{
topTable
}
\description{
helper function for differential binding
}
