% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorFootprints.R
\name{factorFootprints}
\alias{factorFootprints}
\title{plot ATAC-seq footprints infer factor occupancy genome wide}
\usage{
factorFootprints(bamfiles, index = bamfiles, pfm, genome,
  min.score = "95\%", bindingSites, seqlev = paste0("chr", c(1:22, "X",
  "Y")), upstream = 100, downstream = 100)
}
\arguments{
\item{bamfiles}{A vector of characters indicates the file names of bams.}

\item{index}{The names of the index file of the 'BAM' file being processed;
This is given without the '.bai' extension.}

\item{pfm}{A Position frequency Matrix represented as a numeric matrix
with row names A, C, G and T.}

\item{genome}{An object of \link[BSgenome]{BSgenome}.}

\item{min.score}{The minimum score for counting a match.
Can be given as a character string containing a
percentage (e.g. "95%") of the highest possible
score or as a single number.
See \link[Biostrings]{matchPWM}.}

\item{bindingSites}{A object of \link[GenomicRanges]{GRanges} indicates
candidate binding sites (eg. the output of fimo).}

\item{seqlev}{A vector of characters indicates the sequence levels.}

\item{upstream, downstream}{numeric(1) or integer(1).
Upstream and downstream of the binding region for
aggregate ATAC-seq footprint.}
}
\value{
an invisible list of matrixes with the signals for plot.
}
\description{
Aggregate ATAC-seq footprint for a given motif generated
             over binding sites within the genome.
}
\examples{

shiftedBamfile <- system.file("extdata", "GL1.bam",
                             package="ATACseqQC")
library(MotifDb)
CTCF <- query(MotifDb, c("CTCF"))
CTCF <- as.list(CTCF)
library(BSgenome.Hsapiens.UCSC.hg19)
factorFootprints(shiftedBamfile, pfm=CTCF[[1]],
                genome=Hsapiens,
                min.score="95\%", seqlev="chr1",
                upstream=100, downstream=100)

}
\references{
Chen, K., Xi, Y., Pan, X., Li, Z., Kaestner, K., Tyler, J.,
Dent, S., He, X. and Li, W., 2013.
DANPOS: dynamic analysis of nucleosome position and occupancy by sequencing.
Genome research, 23(2), pp.341-351.
}
\author{
Jianhong Ou
}
