% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IGVSnapshot.R
\name{IGVSnapshot}
\alias{IGVSnapshot}
\title{Take IGV snapshots}
\usage{
IGVSnapshot(maxMem, genomeBuild = "hg38", bamFileFullPathOrURLs,
  geneNames = c("GAPDH", "ACTB", "C1orf43", "CHMP2A", "EMC7", "GPI", "PSMB2",
  "PSMB4", "RAB7A", "REEP5", "VCP", "VPS29"), genomicRegions, sessionFile,
  outDir = getwd())
}
\arguments{
\item{maxMem}{A character vector of length 1L containing the maximum usable 
memory for the IGV to be launched, which is defined as the following: 
'mm' - 1.2 GB , 'lm' - 2 GB, 'hm' - 10 GB, ' - 750 MB. For details, please
see the #' parameter memory of function startIGV in the SRAdb package.}

\item{genomeBuild}{A character vector of length 1L: the genome build version.
For example, "hg38" for the human reference genome assembly GRCh38. 
Please see IGV manual for supported genomes of species.}

\item{bamFileFullPathOrURLs}{A character vector containing the full paths 
or ULRs for bam files and matching bam index files. No spaces are allowed 
in the paths or URLs.}

\item{geneNames}{A character vector containing the names of genes or genomic
regions for visualization. For human (maybe other mammalian) ATAC-seq data, 
names of several house-keeping genes are provided as defaults.}

\item{genomicRegions}{A character vector containing genomic regions in the
form of "chr1:20000-40000".}

\item{sessionFile}{A character vector of length 1L containing the name of 
the IGV session file to be saved.}

\item{outDir}{A character vector of length 1L containing the output directory.}
}
\value{
None
}
\description{
Take IGV snapshots for genomic regions of interest.
Prerequisite: Make sure the Integrative Genomics Viewer (IGV) 
(http://software.broadinstitute.org/software/igv/) is installed in a 
user's computer. Be aware of the MAC OSX security and privacy setting. 
To utilize this functionality, be sure that the box for "Enable port", which 
can be found under the toolbar (View -> Preference -> Advanced), is checked 
to allow communication via port.
}
\examples{
if(interactive()){
library(SRAdb)
exampleBams = file.path(system.file('extdata',package='SRAdb'),
             dir(system.file('extdata',package='SRAdb'),pattern='bam$'))
IGVSnapshot(maxMem="mm", genomeBuild="hg18", bamFileFullPathOrURLs= exampleBams,
     geneNames=NULL, genomicRegions="chr1:1-1000",  sessionFile="IGV.session")
}
}
\author{
Haibo Liu
}
