% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NFRscore.R
\name{NFRscore}
\alias{NFRscore}
\title{Nucleosome Free Regions (NFR) score}
\usage{
NFRscore(obj, txs, seqlev = intersect(seqlevels(obj), seqlevels(txs)),
  nucleosomeSize = 150, nucleosomeFreeSize = 100)
}
\arguments{
\item{obj}{an object of \link[GenomicAlignments]{GAlignments}}

\item{txs}{GRanges of transcripts}

\item{seqlev}{A vector of characters indicates the sequence levels.}

\item{nucleosomeSize}{numeric(1) or integer(1). Default is 150}

\item{nucleosomeFreeSize}{numeric(1) or integer(1). Default is 100}
}
\value{
A object of \link[GenomicRanges]{GRanges} with NFR scores
}
\description{
NFR score is calculated for nuclosome signal divided by nuclosome free signal in 
every TSS 400bp windows. 
For each TSS window, it is seperated into 3 parts, first 150 bp for nuclesome 1 (n1) and 
last 150 bp for nucleosome 2 (n2), and the middle 100 bp is for nucleosome free (nf). 
The coverages of 5' ends of the fragments for each part are calculated.
The NFR score is calculated by NFR-score = log2(nf) - log2((n1+n2)/2)
}
\examples{
 
library(GenomicRanges)
bamfile <- system.file("extdata", "GL1.bam", 
                       package="ATACseqQC", mustWork=TRUE)
gal1 <- readBamFile(bamFile=bamfile, tag=character(0), 
                    which=GRanges("chr1", IRanges(1, 1e6)), 
                    asMates=FALSE)
library(TxDb.Hsapiens.UCSC.hg19.knownGene)
txs <- transcripts(TxDb.Hsapiens.UCSC.hg19.knownGene)
nfr <- NFRscore(gal1, txs)
}
\author{
Jianhong Ou
}
