% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakDet.R
\name{peakdet}
\alias{peakdet}
\title{Detect peak positions}
\usage{
peakdet(y, delta = 0, silence = TRUE)
}
\arguments{
\item{y}{A vector of numeric where to search peaks}

\item{delta}{A numeric of length 1, defining the local threshold
for peak detection. If it is set to 0, the delta will be set to
1/10 of the range of y.}

\item{silence}{logical(1). 
If false, echo the delta value when delta is set as 0.}
}
\value{
A list with peakpos and valleypos. Both peakpos and valleypos are
vectors of numeric which indicate the positions of peak or valley.
}
\description{
Detect the peaks positions and valley positions. The algorithm is modified
from github::dgromer/peakdet
}
