% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footprintsScanner.R
\name{prepareBindingSitesList}
\alias{prepareBindingSitesList}
\title{helper function for preparing the binding list}
\usage{
prepareBindingSitesList(pfms, genome, min.score = "85\%", maxSiteNum = 5000)
}
\arguments{
\item{pfms}{A list of Position frequency Matrix represented as a numeric matrix
with row names A, C, G and T.}

\item{genome}{An object of \link[BSgenome]{BSgenome}.}

\item{min.score}{The minimum score for counting a match.
Can be given as a character string containing a
percentage (e.g. "95%") of the highest possible
score or as a single number.
See \link[Biostrings]{matchPWM}.}

\item{maxSiteNum}{numeric(1). Maximal number of predicted binding sites.
if predicted binding sites is more than this number, top maxSiteNum binding
sites will be used.}
}
\description{
helper function for preparing the binding list
}
\examples{
library(MotifDb)
motifs <- query(MotifDb, c("Hsapiens"))
motifs <- as.list(motifs)
library(BSgenome.Hsapiens.UCSC.hg19)
#bindingSitesList <- prepareBindingSitesList(motifs, genome=Hsapiens)
}
