% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwmscores.R
\name{pwmscores}
\alias{pwmscores}
\title{max PWM scores for sequences}
\usage{
pwmscores(pwm, subject)
}
\arguments{
\item{pwm}{A Position Weight Matrix represented as a numeric matrix
with row names A, C, G and T.}

\item{subject}{Typically a \link[Biostrings]{DNAString} object.
A \link[IRanges]{Views} objecton
a \link[Biostrings]{DNAString} subject,
a \link[Biostrings]{MaskedDNAString} object,
or a single character string, are also supported.
IUPAC ambiguity letters in \code{subject} are ignored (i.e.
assigned weight 0) with a warning.}
}
\value{
a numeric vector
}
\description{
calculate the maximal PWM scores for each given sequences
}
\author{
Jianhong
}
