% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readsDupFreq.R
\name{readsDupFreq}
\alias{readsDupFreq}
\title{reads duplicates frequency}
\usage{
readsDupFreq(bamFile, index = bamFile)
}
\arguments{
\item{bamFile}{character. bam file name.}

\item{index}{character. name of index file.}
}
\value{
a 2 column matrix with reads duplicates frequency
}
\description{
Calculate the frequency of read duplication based on alignment coordinates
determined by rname, strand, pos, cigar, mrnm, and isize.
}
\examples{
bamFile <- system.file("extdata", "GL1.bam", package = "ATACseqQC")
freq <- readsDupFreq(bamFile)
}
\author{
Haibo Liu
}
