% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saturationPlot.R
\name{saturationPlot}
\alias{saturationPlot}
\title{Saturation plot}
\usage{
saturationPlot(subsamplingPeakFiles, subsamplingSizes, peakCaller = "MACS2",
  sep = "\\t", header = c(TRUE, FALSE), fdr = 0.05, fdrCol,
  startCol = 2, endCol = 3, skipLines, ...)
}
\arguments{
\item{subsamplingPeakFiles}{a list of files from peak calling tools}

\item{subsamplingSizes}{a vector of integers, which are the sizes of subsamples for 
peak calling. subsamplingPeakFiles should match subsamplingSizes
in a correct order.}

\item{peakCaller}{"MACS2", "ZINBA", "SPP", or "F-seq". Current only support MACS2 output}

\item{sep}{a string, column separator}

\item{header}{a boolean (TRUE or FASLE), showing whether there are headers in the 
peak files}

\item{fdr}{decimal between 0 and 1, a cutoff of statistical significance of peak detection}

\item{fdrCol}{integer, column index for fdr}

\item{startCol}{integer, column index for start positions of peak regions}

\item{endCol}{integer, column index for end positions of peak regions}

\item{skipLines}{integer, the number of lines (comment or instruction) to skip when peak 
files are read into R}
}
\description{
plot the saturations
}
\examples{
if(interactive()){
}
}
\author{
Haibo Liu
}
