% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiftReads.R
\name{shiftReads}
\alias{shiftReads}
\title{shift read for 5'end}
\usage{
shiftReads(x, positive = 4L, negative = 5L)
}
\arguments{
\item{x}{an object of GAlignments}

\item{positive}{integer(1). the size to be shift for positive strand}

\item{negative}{integer(1). the size to be shift for negative strand}
}
\value{
an object of GAlignments
}
\description{
shift reads for 5'ends
}
\author{
Jianhong Ou
}
