% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footprintsScanner.R
\name{footprintsScanner}
\alias{footprintsScanner}
\title{scan ATAC-seq footprints infer factor occupancy genome wide}
\usage{
footprintsScanner(bamfiles, index = bamfiles, bindingSitesList,
  seqlev = "chr1", upstream = 100, downstream = 100)
}
\arguments{
\item{bamfiles}{A vector of characters indicates the file names of bams. 
All the bamfiles will be pulled together.}

\item{index}{The names of the index file of the 'BAM' file being processed;
This is given without the '.bai' extension.}

\item{bindingSitesList}{A object of \link[GenomicRanges:GRangesList-class]{GRangesList} indicates
candidate binding sites (eg. the output of fimo).}

\item{seqlev}{A vector of characters indicates the sequence levels.}

\item{upstream, downstream}{numeric(1) or integer(1).
Upstream and downstream of the binding region for
aggregate ATAC-seq footprint.}
}
\value{
an invisible list of matrixes with the signals for plot.
It includes:
 - signal    mean values of coverage for positive strand and negative strand
             in feature regions
 - spearman.correlation    spearman correlations of cleavage counts in the 
             highest 10-nucleotide-window and binding prediction score.
 - bindingSites    predicted binding sites.
}
\description{
Aggregate ATAC-seq footprint for a bunch of motifs generated
             over binding sites within the genome.
}
\examples{

bamfile <- system.file("extdata", "GL1.bam",
                       package="ATACseqQC")
bsl <- system.file("extdata", "jolma2013.motifs.bindingList.95.rds",
                  package="ATACseqQC")
bindingSitesList <- readRDS(bsl)
footprintsScanner(bamfile, bindingSitesList=bindingSitesList)

}
\references{
Chen, K., Xi, Y., Pan, X., Li, Z., Kaestner, K., Tyler, J.,
Dent, S., He, X. and Li, W., 2013.
DANPOS: dynamic analysis of nucleosome position and occupancy by sequencing.
Genome research, 23(2), pp.341-351.
}
\author{
Jianhong Ou
}
