% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragSizeDist.R
\name{fragSizeDist}
\alias{fragSizeDist}
\title{fragment size distribution}
\usage{
fragSizeDist(bamFiles, bamFiles.labels, ylim = NULL, logYlim = NULL)
}
\arguments{
\item{bamFiles}{A vector of characters indicates the file names of bams.}

\item{bamFiles.labels}{labels of the bam files, used for pdf file naming.}

\item{ylim}{numeric(2). ylim of the histogram.}

\item{logYlim}{numeric(2). ylim of log-transformed histogram for the insert.}
}
\value{
Invisible fragment length distribution list.
}
\description{
estimate the fragment size of bams
}
\examples{
bamFiles <- system.file("extdata", "GL1.bam", package="ATACseqQC")
bamFiles.labels <- "GL1"
fragSizeDist(bamFiles, bamFiles.labels,
             ylim=c(0, 1e4), logYlim=log10(c(5e-3, 2)))
}
\author{
Jianhong Ou
}
