% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footprintsScanner.R
\name{prepareBindingSitesList}
\alias{prepareBindingSitesList}
\title{helper function for preparing the binding list}
\usage{
prepareBindingSitesList(pfms, genome, seqlev = paste0("chr", c(1:22, "X",
  "Y")), expSiteNum = 5000)
}
\arguments{
\item{pfms}{A list of Position frequency Matrix represented as a numeric matrix
with row names A, C, G and T.}

\item{genome}{An object of \link[BSgenome:BSgenome-class]{BSgenome}.}

\item{seqlev}{A vector of characters indicates the sequence levels.}

\item{expSiteNum}{numeric(1). Expect number of predicted binding sites.
if predicted binding sites is more than this number, top expSiteNum binding
sites will be used.}
}
\description{
helper function for preparing the binding list
}
\examples{
library(MotifDb)
motifs <- query(MotifDb, c("Hsapiens"))
motifs <- as.list(motifs)
library(BSgenome.Hsapiens.UCSC.hg19)
#bindingSitesList <- prepareBindingSitesList(motifs, genome=Hsapiens)
}
